Randomized item with weight

+++ Assets/Core/Scripts/ScriptableObject Script/ItemChanceBase.cs
@@ -5,13 +5,13 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Item", menuName = "Chance/Create new Basic Chance")]
 public class ItemChanceBase : ScriptableObject
 {
-    [SerializeField] float common;
-    [SerializeField] float uncommon;
-    [SerializeField] float rare;
-    [SerializeField] float legendary;
+    [SerializeField] int common;
+    [SerializeField] int uncommon;
+    [SerializeField] int rare;
+    [SerializeField] int legendary;
 
-    public float GetCommonChance() { return common; }
-    public float GetUncommonChance() { return uncommon; }
-    public float GetRareChance() { return rare; }
-    public float GetLegendaryChance() { return legendary; }
+    public int GetCommonChance() { return common; }
+    public int GetUncommonChance() { return uncommon; }
+    public int GetRareChance() { return rare; }
+    public int GetLegendaryChance() { return legendary; }
 }
